+++ tapo/src/api/discovery.rs
@@ -2,10 +2,13 @@ mod aes_discovery_query_generator;
 mod device_discovery;
 mod device_discovery_raw;
 mod device_type;
+mod discovery_raw_result;
 mod discovery_result;
 
 pub use device_discovery::*;
 #[cfg(feature = "debug")]
 pub use device_discovery_raw::*;
 pub use device_type::*;
+#[cfg(feature = "debug")]
+pub use discovery_raw_result::*;
 pub use discovery_result::*;